requirementts.txt
streamlit
google-generativeai==0.15.2
pillow
request(optional)